ServiceProvider1/requirements.txt
absl-py==0.13.0
aioredis==1.3.1
aiosqlite==0.16.1
amqp==5.0.6
asgiref==3.3.4
astroid==2.5.1
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
Babel==2.9.1
bcrypt==3.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
cachetools==4.2.2
celery==5.1.2
certifi==2020.12.5
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.0
chardet==4.0.0
chatbot==1.5.2b0
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
comtypes==1.1.9
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cycler==0.10.0
daphne==3.0.2
defusedxml==0.7.1
diff-match-patch==20200713
distro==1.6.0
dj-database-url==0.5.0
Django==3.2.4
django-heroku==0.3.1
django-import-export==2.5.0
django-templated-mail==1.1.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djongo==1.3.6
djoser==2.1.0
docker==5.0.0
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
emoji==1.2.0
et-xmlfile==1.1.0
factor-analyzer==0.3.1
Flask==2.0.1
flatbuffers==1.12
gast==0.4.0
google-auth==1.32.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
gunicorn==20.1.0
h5py==3.1.0
heroku==0.1.4
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
iso8601==0.1.16
isort==5.7.0
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.1
joblib==1.0.1
Js2Py==0.71
jsonschema==3.2.0
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
kombu==5.1.0
lazy-object-proxy==1.5.2
Markdown==3.3.4
MarkupPy==1.14
MarkupSafe==2.0.1
mathparse==0.1.2
matplotlib==3.4.2
mccabe==0.6.1
model-selection==0.0.1
msgpack==1.0.2
mysql-connector==2.2.9
mysql-connector-python==8.0.24
natsort==7.1.1
nltk==3.6.2
numpy==1.19.5
oauthlib==3.1.1
odfpy==1.4.1
opencv-python==4.5.2.54
openpyxl==3.0.7
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.1
paramiko==2.7.2
Pillow==8.3.1
Pint==0.17
pipwin==0.5.1
prompt-toolkit==3.0.19
protobuf==3.15.8
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///C:/Users/nidhi/pipwin/PyAudio-0.2.11-cp39-cp39-win_amd64.whl
pycparser==2.20
pygame==2.0.1
PyHamcrest==2.0.2
pyjsparser==2.7.1
PyJWT==2.1.0
pylint==2.7.2
pymongo==3.11.4
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyPika==0.48.8
pypika-tortoise==0.1.1
pypiwin32==223
PyPrind==2.11.3
pyrsistent==0.18.0
pySmartDL==1.3.4
python-dateutil==1.5
python-dotenv==0.19.0
python3-openid==3.2.0
pyttsx3==2.90
pytz==2021.1
pywin32==227
PyYAML==3.13
redis==3.5.3
regex==2021.7.6
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.0
seaborn==0.11.1
seedir==0.1.4
service-identity==21.1.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
SQLAlchemy==1.2.19
sqlparse==0.2.4
tablib==3.0.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
termcolor==1.1.0
texttable==1.6.4
threaded==4.1.0
threadpoolctl==2.2.0
tk==0.1.0
tkcalendar==1.6.1
toml==0.10.2
tortoise-orm==0.17.6
tqdm==4.61.2
Twisted==21.7.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typing-extensions==3.7.4.3
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.4
varclushi==0.1.0
vine==5.0.0
wcwidth==0.2.5
websocket-client==0.59.0
Werkzeug==2.0.1
wget==3.2
whitenoise==5.3.0
wikipedia==1.4.0
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.4.0
